moved the tests back to the existing test files; added them in a new class
